# odd-platform-specification/openapi.yaml
openapi: 3.0.3
info:
  title: ProspectLog data catalog HTTP API contract
  description: ProspectLog data catalog HTTP API contract
  contact:
    email: ndementev@provectus.com
    url: https://provectus.com
    name: Nikita Dementev
  version: 0.0.1
servers:
  - url: 'http://localhost'
    description: stub
...
paths:
...
  /api/owners:
    get:
      summary: List of owners
      description: Gets the list of existing owners
      operationId: getOwnerList
      parameters:
        - $ref: './components.yaml/#/components/parameters/SearchParam'
        - $ref: './components.yaml/#/components/parameters/PageParam'
        - $ref: './components.yaml/#/components/parameters/SizeParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/OwnerList'
      tags:
        - owner
    post:
      summary: Create owner
      operationId: createOwner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/components/schemas/OwnerFormData'
      responses:
        '201':
          description: The resource has been successfully created
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/Owner'
      tags:
        - owner
  /api/owners/{owner_id}:
    put:
      summary: Update owner
      description: Updates existing owner
      operationId: updateOwner
      parameters:
        - name: owner_id
          in: path
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/components/schemas/OwnerFormData'
      responses:
        '201':
          description: The resource has been successfully updated
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/Owner'
      tags:
        - owner
    delete:
      summary: Delete owner
      description: Deletes existing owner
      operationId: deleteOwner
      parameters:
        - name: owner_id
          in: path
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '204':
          $ref: './components.yaml/#/components/responses/Deleted'
      tags:
        - owner
...
  /api/tags:
    get:
      summary: List of popular tags
      description: Gets the list of existing tags sorted by popularity
      operationId: getPopularTagList
      parameters:
        - $ref: './components.yaml/#/components/parameters/PageParam'
        - $ref: './components.yaml/#/components/parameters/SizeParam'
        - $ref: './components.yaml/#/components/parameters/SearchParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/TagsResponse'
      tags:
        - tag
    post:
      summary: Create a tag
      description: Creates a tag
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/components/schemas/BulkTagFormData'
      responses:
        '201':
          description: The resource has been successfully created
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/TagList'
      tags:
        - tag
  /api/tags/{tag_id}:
    put:
      summary: Update tag
      description: Updates existing tag
      operationId: updateTag
      parameters:
        - name: tag_id
          in: path
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/components/schemas/TagFormData'
      responses:
        '201':
          description: The resource has been successfully updated
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/Tag'
      tags:
        - tag
    delete:
      summary: Delete tag
      description: Deletes existing tag
      operationId: deleteTag
      parameters:
        - name: tag_id
          in: path
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '204':
          $ref: './components.yaml/#/components/responses/Deleted'
      tags:
        - tag
...
  /api/search/{search_id}:
    get:
      summary: Get search facets by search ID
      description: Gets search facets by parameter which can be found in search endpoint response
      operationId: getSearchFacetList
      parameters:
        - name: search_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/SearchFacetsData'
      tags:
        - search
    put:
      summary: Update search facets by search ID
      description: Updates search facets of a certain search ID by provided search parameters
      operationId: updateSearchFacets
      parameters:
        - name: search_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/components/schemas/SearchFormData'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/SearchFacetsData'
      tags:
        - search
  /api/search/{search_id}/facet/{facet_type}:
    get:
      summary: Get countable filters for chosen facet type
      description: Gets countable filters for chosen facet type
      operationId: getFiltersForFacet
      parameters:
        - name: search_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: facet_type
          in: path
          required: true
          schema:
            $ref: './components.yaml/#/components/schemas/MultipleFacetType'
        - $ref: './components.yaml/#/components/parameters/PageParam'
        - $ref: './components.yaml/#/components/parameters/SizeParam'
        - $ref: './components.yaml/#/components/parameters/SearchParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './components.yaml/#/components/schemas/CountableSearchFilter'
      tags:
        - search
  /api/search/{search_id}/results:
    get:
      summary: Get search results by search ID
      description: Returns paginated list of search results for a given search ID
      operationId: getSearchResults
      parameters:
        - name: search_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: './components.yaml/#/components/parameters/PageParam'
        - $ref: './components.yaml/#/components/parameters/SizeParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/DataEntityList'
      tags:
        - search
...
  /api/metadata/fields:
    get:
      summary: Get custom metadata fields
      description: Gets all available custom metadata fields
      operationId: getMetadataFieldList
      parameters:
        - $ref: './components.yaml/#/components/parameters/SearchParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/MetadataFieldList'

      tags:
        - metadata
    post:
      summary: Create new custom metadata field
      description: Creates new custom metadata field
      operationId: createCustomField
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/components/schemas/MetadataFieldFormData'
      responses:
        '201':
          description: The resource has been successfully created
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/MetadataField'
      tags:
        - metadata
  /api/metadata/fields/{metadata_field_id}:
    put:
      summary: Update custom metadata field
      description: Updates custom metadata field
      operationId: updateCustomField
      parameters:
        - $ref: './components.yaml/#/components/parameters/DataEntityMetadataFieldIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/components/schemas/MetadataUpdateCustomFieldFormData'
      responses:
        '201':
          description: The resource has been successfully updated
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/MetadataField'
      tags:
        - metadata
    delete:
      summary: Delete custom metadata field
      description: Deletes custom metadata field
      operationId: deleteCustomField
      parameters:
        - $ref: './components.yaml/#/components/parameters/DataEntityMetadataFieldIdParam'
      responses:
        '204':
          $ref: './components.yaml/#/components/responses/Deleted'
      tags:
        - metadata
...